metrics: Add support for different aggregation functions
The metrics layer supports aggregate metrics by labels when
reporting them. This patch introduces the ability to select the
aggregation function that will be used.

The default aggregation function is "sum," which matches the majority of
cases. However, it is now possible to use "avg," "min," and "max" as
well.

In addition to adding the enum for selecting a function, a comparable
function (which will be used by "min" and "max") and an "avg" function
were also added.

With this patch, when marking a metric for aggregation, it will be
possible to set which aggregation function will be used.

Usage example

ms::make_gauge("messages", ms::description("Average messages"), _msg)
  .aggregate({ms::shard_label}, ms::aggregate_function_type::AVG)

Signed-off-by: Amnon Heiman <amnon@scylladb.com>