future: de-variadicate make_ready_future() and similar helpers
The functions

   make_ready_future
   make_exception_future
   current_exception_as_future

Are all variadic, but since we eliminated variadic futures they
can be made ordinary non-variadic templates. Do so to clean up
the code a little. "T..." is replaced by "T = void" to preserve
current behavior with empty parameters.

In one case, a forward declaration had to be added. This is because
the function gained a default template parameter, but was also
previously declared as a friend. Since friends cannot define
default template parameters, we add a forward declaration.

Closes #1593