condition-variable: replace the coroutine wakeup task with a promise
condition_variable::awaiter implements the coroutine resumption after
signal() by scheduling itself as a task which resumes the coroutine. The
awaiter itself is (usually) allocated inside the coroutine frame, and
dies after its containing co_await expression is resumed.

This is a problem, because the reactor expects tasks to stay alive while
they are running — it saves a pointer to the current task in
reactor::_current_task, which is used to find the parent task when
printing the tasktrace. Because the awaiter task dies immediately after
returning control to its parent, reactor::_current_task is left pointing
to some garbage in the coroutine frame and current_tasktrace() will
segfault when used before the next task switch.

This patch replaces this wakeup logic with a promise-future pair.
Instead of scheduling a handle resume task, signal() just triggers a
promise. All coroutine logic is reused from the implementation of
co_await for futures.

Fixes #1599