fstream: Use private file API
The fstream code needs to specify the priority class to call file IO
methods with. This is because IO priority fstream works with is
specified on its options and have to be explicit. However, file public
methods that accept explicit priority class argument are going to be
deprecated and removed, so the fstream need to use some other way to
pass priority class to file. The best option here seems to be the
file::..._impl() methods.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>