metrics_types.hh: add optional native histogram information
Prometheus native histograms are an alternative dense representation of
a histogram.

With a native histogram, bucket size cannot be arbitrary. Instead, each
bucket is 2^(2^(-schema)) bigger than the previous one, where schema is
an integer number.

For example, if the schema is 0, each bucket will be 2 times bigger than
the previous one.

This patch adds an optional native histogram info to the metrics
histogram data type.

native_histogram_info holds the schema and min_id.

If the parameter is present, the Prometheus layer may choose to send
the histogram in a native histogram format.