io_queue: Maintain per-shard-class stall threshold
The stall threshold is the duration a request is expected to get
dispatched and executed within. In current model it's expected to be
io-latency-goal value, but it's better to have it configurable.

When the threshold is broken, the report helper is called and the
threshold is doubled. Once in a while the threshold is lowered to
allow for more reports in the future.

Reported is the number of queued requests in the whole queue and
per-class as well as the number of currently executing requests also
both kinds -- in the whole queue and per-class.

fixes: #1311

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>