reactor: stop _aio_eventfd loop also on shard other than 0
Fixes a regression introduced in f6a2c0125f.

_aio_eventfd loop stops when it's woken up with _stopping set.
Thus stop_aio_eventfd_loop(), whose job is to stop the loop,
has to be preceded by setting _stopping to true.

In f6a2c0125f the assignment to _stopping was moved to a different place,
which accidentally caused it to happen only on shard 0, while
it has to happen on all shards.
(Because if it doesn't happen on some shard, this shard will hang on exit.)

Fixes #1610