program_options: Introduce updateable_value<>
The updateable_value is like value, but it should be initialized with a
noncopyable_function that would be called when the option is updated
run-time.

The updating function is called in two cases:

- The option is set in config file. In that case it's up to the callback
  to check if the option changed its exact value or not
- The option is not set in config file and it's currently not defaulted.
  This means "user wants it back to default" and the same -- its up to
  the callback to check if the default value differs from current one.

Additionally, the moveable lock object is passed to the caller. While
the value is "locked" further HUPs on it will be ignored. Updating
function must release the object when it's ready to be called again.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>