shared_future: remove support for variadic futures
shared_future is a variadic template for two reasons: first,
we used to support variadic futures. Second, it can be tuned
using a with_clock<> initial template parameter. We removed
support for variadic futures, but due to the second reason
shared_future should remain variadic. We could change it
to have two parameters, but the role of each positional parameter
changes according to what arguments are provided, which is awkward.

So, we just drop support for variadic futures while keeping
shared_future itself variadic. The whole thing is concentrated
in a helper future_option_traits which interprets the parameters
using specializations. All that is needed is to adjust the
specializations.

There's a bit of extra complexity since shared_future synthesizes
a tuple type from its parameters, but this doesn't work for
void (e.g. tuple<void> is bad). Fortunately there is
future<T>::tuple_type that provides the same thing.

To clean up the whole thing, we drop the clever
future_option_traits::parametrize template, and replace it
with its two instantiations (one for future, one for promise).
This simplifies things considerably.

Closes #1616