rpc server: Warn server stop is stuck after 1 minute
When server stops it waits for several internal futures to resolve, and
one of them -- the gate that picks up all response sendings to finish.
However, this gate also covers the RPC handler callbacks, which are
external to the RPC code. So if the handler gets stuck it would be good
to know this to tell internal RPC problems from external ones.

For that on server::stop() arm a timer for 1 minute then go waiting for
the stopping futures to resolve then cancel the timer. In timer handler
show the message in seastar logger and print the amount of in-flight
handlers.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>