memory: Limit alloc_sites size
Basic check to prevent alloc_sites from growing too large.

Note that for simplicity reasons the current implemenation still behaves
as if we sampled otherwise (i.e.: stores alloc site with a nullptr).

The 1000 limit is currently kinda arbitrary to bound memory overhead. If
running with very low sample rates we'd probably want to make this a
config value/define.