file_impl: Introduce maybe_priority_class_ref
All file_impl inheritants are going to have two implementations -- one
legacy and one for the +1 API level. The difference is in the way the
io_priority_class argument is passed around. In the legacy case it's the
explicit const reference, in the latter case the priority_class won't
even exist. However, _implementations_ of the methods do manipulate the
"pc" argument internally and it's not nice to code them twice or wrap
internals of the methods with ifdefs.

So this patch wraps all overriden methods of file_impl inheritants into
private calls that accept maybe_priority_class_ref argument. The
overriden methods stay legacy by accepting const io_priority_class& one
and constructing the maybe-ref from it.

Also, to make the patch compile, recently introduced priority_class for
io-queue level gains the ability to be constructed directly from the
maybe-ref thing.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>