file: Use maybe_priority_class_ref
A difficulty described in the previous patch also exists in a light form
at the file-level, so it makes sense to patch it to use the maybe-ref as
well.

At this point the file methods that instantiate default_priority_class()
behind the scenes construct the maybe-ref with no arguments. For now
this means "reference default priority class".

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>