file: Implement API for API level 7
It's not yet declared, but implementing file API for it in advance makes
the patch with this level much shorter and earier to review.

In the new level the file API changes in two ways
- IO priority class argument goes away
- io_intent* arguments is mandatory for _impl-s

First about the 2nd change. When the io_intent was introduced the
file_impl got default implementation for the methods accepting the
intent pointer -- it called "legacy" methods dropping the intent thus
implementing the idea -- if the foo_file_impl wants to support requests
cancellation, it has to do it voluntarily and explicitly. Now it's gone,
all foo_file_impl will need to ignore the intent explicitly if they
don't want to support it.

Next, regarding the 1st change. The idea is that the priority_class
doesn't exist in the file layer. By the time request gets to the
io_queue::queue_request() the latter would simply pick up the current
scheduling group and will work with it. Previous patches prepare the
ground for this change by introducing the maybe_priority_class_ref
thing. This patch implements the maybe-ref for the next API level by
making it an empty struct. This lets file_impl-s pass it transparently
down to the io-queue level, the latter would try to construct its
internal::priority_class instance from it and will pick current sched
class.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>