reactor: Update and rename queues for API level 7
A "continuation" of the previous patch. When sched group is renamed or
needs to change its shares, all queues that managed to pick up name and
shares before it needs to be updated. Not a major change, current IO
priority classes renaming and shares-changing code does so, this patch
just kicks the io-queues from another API call.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>