build: only pass CMAKE_C_COMPILER when building ingredients
Seastar is a C++ library, and we don't use C compiler when building
the tree. But we could use the C compiler when building some
ingredients. so, in this change, we only pass `-DCMAKE_C_COMPILER=..`
option to CMake when building ingredents. this change should silence
the warning of

```
CMake Warning:
  Manually-specified variables were not used by the project:

    CMAKE_C_COMPILER
```

this warning does not hurt, but it could be misleading.

also, in this change, the f-string is used for better readability.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>