file: s/(void)/()/ in function's parameter list
according to the C++ standard draft $9.3.4.6 [dcl.fct], see
https://eel.is/c++draft/dcl.fct#4 .

> A parameter list consisting of a single unnamed parameter of
> non-dependent type void is equivalent to an empty parameter list.

and the symbol of a C++ function of `int foo(void)` is indeed
identical to that of `int foo()`: both of them are `_Z3foov`. so,
there is no need to keep the void parameter in the parameter list.
this is a good practice in the C language though, the `void` in the
parameter sends the message explicitly that: there is no parameters
here.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>