rpc: add `server&` and `connection_id` to `client_info`
`rpc::client_info` instance lives inside a `server_connection` which is
managed by a `server`. Add a reference to that `server` and the
`connection`'s ID inside `client_info`.

This will be useful for implementing RPC handlers that want to do
something to the connection or save some information about it for later
usage.

In Scylla we'll use it to store a mapping
`host ID -> connection ref`
which we'll then use to abort all connections to a given host if we
decide to ban it from the cluster.