http/client: Count and limit the number of connections
There can be two kinds of connections for a client -- active and idle.
The former is the connection created from the factory or withdrawn from
pool that's currently surving a request. The latter is the connection
that's listed into the pool.

To count the total number of connections the previously introduced RAII
backref is used -- it increments the counter on creation and decrements
it on destruction.

When the pool is empty the client needs to check if it can spawn more
connections and, if not, wait for one of the events. It's either an
existing active connection gets back to pool or any of the existing
connections gets closed and releases a "slot". So, the waiting, if it
happens, is aborted by the backref destructor or by the client itself
when it puts back a connection into the pool.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>