util: Import Android Memory Sampler
Imports the sampler.h file from the Android Open Source Project from:

https://cs.android.com/android/platform/superproject/+/013901367630d3ec71c9f2bb3f3077bd11585301:external/perfetto/src/profiling/memory/sampler.h

In this commit the file is imported as is. The adaption to seastar and
our needs is done in a following commit. This is so that it stays clear
what was originally imported and what our changes are.