net,rpc: Turn on batch-flushes for RPC sockets only
When calling connected_socket::output() it unconditionally turns on
batch flushes for the created output_stream. However, it's only
meaningful for RPC sockets whose users may put several independent
messages into it and wait for responce.

Contrary to this http client sockets work in pure ping-pong model --
send request, wait for reply. For those turning batch-flush makes
things worse, because waiting for the reply is guaranteed not to get
one until the upcoming reactor tick and batch-flush poller loop. This is
2x worse when 'continue:' header is in use, because sending a request
involves flushing the request in the middle and waiting for the server
confirmation to continue.

The proposal is to make batch-flush OFF by default and teach RPC layer
to configure its socket's output stream with it explicitly.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>