log: report scheduling group along with shard id
The scheduling group provides useful information like which tenant is
currently executing, and is inexpensive to provide, so log it with every
message. because the names of scheduling group could be too long
to be printed in every single logging message, a shortened version is
used. the shortname is specified when creating a scheduling group or
when renaming it. so the involved functions are updated so new variants
is added, these variants allows user to specify an optional shortname
for the scheduling group. if user continue using the old API, the
shortname is created automatically with two strategies

* if the name contains underscore, the shortened name is translated
  using following pseudo code:
  ```py
  ''.join(part[0] for part in name.split("_", maxsplit=4))
  ```
  for instance, the short name of "memtable_to_cache" would be "mc"
* if the name does not contain underscore, the shortened name would
  be the first 4 characters of the full name, so the short name of
  "streaming"  would be "str".

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>