HTTP headers in http_response are now case insensitive
Previously to this commit only the headers in the http requests
were case insensitive. This failed to handled a scenario in which a
server returns headers with different cases.
The most impactful case is `content-type` instead of `Content-Type`.
In case a server uses lower case, seastar would ignore the content and
return an empty reply.

To address this issue the same case-insensitive logic was reused.
Move the case_insensitive_cmp/case_insensitive_hash to a utility module.

Closes #1672