http/client: Add ability to change pool size on the fly
The client::set_maximum_connections() method changes the limit.

If the new value is greater than the old one, it needs to kick potential
get_connection() sleeprs so that they could create new connections.

If the new value is less than the old one, the excessive connections are
closed. First, the pool is shrunk. If that's not enough the method waits
for active connections to close. For that, when an active connection is
put and the client is being shrunk, the connection doesn't get into the
pool but is closed instantly.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>