response_frame: Use constants instead of magic numbers
The response is sent in two steps. First templatized helper allocates a
buffer, offsets sizeof(header) bytes and marshals the message into it.
Then server generates header into the reserved space. Both places use
magic hard-coded numbers that should be in-sync with each other.

It's better to keep the headroom size as named constant and make the
rest of the code use/check it

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>