rpc: Exchange handler duration with server responses
Introduce new negotiation parameter that tells if the client wants
server to send it back the handler execution duration. This is for
compatibility with older clients that expect server reply to be of some
fixed (shorter) size.

New response frame, if negotiated, includes the 32 bit handler duration
value in microseconds right next to the existing bits. If no duration
was provided (sending error response, for example), the value is 0xF..F.
After parsing, the duration is returned as optional value and 0xF..F
maps to disengaged one.

For now client makes no use of it, next patch will.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>