perf: add sstring_perf for testing sstring performance
it turns out, we are using basic_sstring<int8_t, ...> in our project.
but sstring uses std::char_traits<char_type>::compare() to compare the
strings, but char_traits<char_type>::compare() only use memcmp() for
char and wchar_t specliazations. for other char_types, it just does
the comparison lexicographically with a loop. this is far from
optimised in comparison to memcmp().

so add a test case to observe this behavior.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>