reactor: use s/::free()/::io_uring_free_probe()/
use ::io_uring_free_probe() instead of ::free() to free the memory
chunk pointed by `probe`. io_uring allocates the memory chunk by
itself. so there are chances that the precompiled liburing is using
the vanilla malloc from the libc, while we free the memory chunk
using the overridden version which works with the customized allocator.

in other words, we cannot assume which malloc() was used. in this
change, ::io_uring_free_probe() is used, as this function is provided
by liburing, so we can ensure that the free() used by it always matches
with the malloc() called by io_uring_get_probe_ring().

this function was introduced since liburing 2.0, so we are safe, as
Seastar's minimal required liburing is 2.0.

this should address the issue reported by ASan when the tree is compiled
with liburing 2.4, like:

AddressSanitizer: attempting free on address which was not malloc()-ed

Fixes #1692
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>