smp: warn if --memory parameter is not supported
This parameter is ignored if SEASTAR_DEFAULT_ALLOCATOR
build option is set. For Scylla this is true
for debug builds to enable ASAN allocator.
There was no clear indication that it was
not applied before, which could cause
misunderstanding and wasted time during development.

--memory is one of the options in
smp_options and reactor_options
which won't have effect if Seastar allocator
was not compiled in. We emit a warning
only for --memory since other options are
way less common and impactful.