smp: introduce memory prefaulter
With --lock-memory, we use mlock() to make sure memory isn't swapped
out, as that could induce latency spikes. However, we still see
latency spikes when the memory is first faulted in. We use MCL_FUTURE,
since locking all memory immediately results in huge startup delay,
as this work is single-threaded.

To avoid those latency spikes, we add new threads that fault the
memory in, locking it. To cause minimum disruption, we create just
one thread per NUMA node and place it in the SCHED_IDLE scheduling
policy.

We round-robin between the shards of a single NUMA node. Within a shard,
we fault at both the low bound and the upper bound of the address space.
This helps applications that manage two heaps that grow towards each
other (ScyllaDB).

Since memory_prefaulter is forward-declared in its container class smp,
we have to deinline smp::~smp() so its definition sees memory_prefaulter's
destructor.