http/client: Add "total new connections" metrics
The metrics is the monotonic counter that gets ++ed every time the
factory is kicked for new connection.

The recommended usage is to get the time-derivative from the counter and
check _this_ value. If it's large, it means that the existing
connections are not getting back to pool due to connection errors or
unexpected server status codes. This, in turn, can be used to explain
large delays in requests, as each time a request is run over a new
connection it doesn't benefit from connections keep-alive and gets extra
latency.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>

Closes #1718