memcached: do not used deprecated API
in 8038c131b7c953319dd69d045c7c5b71d34a6018, lw_shared_ptr operator=(T&&)
was deprecated. but since we don't build apps when performing CI or
general testing. this caller was not updated in that very change.

when compiling the whole tree, this deprecated warning was identified.
so let's fix it.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #1720