reactor: epoll backend: initialize _highres_timer_pending
The variable is subject to a race condiotion and can be read before use.

This is a benign race; in the worst case there can be a single
false-positive in the lifetime of the reactor, and the false positive's
impact is a few wasted cycles trying to expire a highres timer where
in fact none are ready. In addition it's in the epoll backend that
noone uses.

However, it occasionally fails some debug mode tests that have strict
ubsan checks, so it's worth fixing.

Closes #1721