build: serialize calls to openssl certificate generation
During build, we generate a few SSL certificates for testing. The
generation process as a .srl file containing a serial number, which
is both read and written. It appears openssl doesn't bother to lock
it, so if ninja happens to run two such processes concurrently, one
of them can fail when reading a partially written file.

Fix this by placing these build steps into a job pool with a
concurrency of one.

Also add the implicit dependency into the command declarations
this doesn't help much, since the dependency is an input/output
file which can't be expressed by the build system, but helps
make the file visible.

Closes #1725