rpc: add zstd streaming compressor
TODO:

- Add zstd to seastar dependencies.
- Check that exceptions in the constructor of the compressor are
  handled properly.

Considerations:

With window size 16 (chosen in this patch), zstd will be able to
backreference 64 kiB into the past, at the cost of ~521 kiB per
compressor and 413 kiB per decompresor, which adds up to 1 MiB
per connection which uses this compression.
(Note: since these are separate allocations now, it really adds up
to 1.5 MiB due to separate rounding. But we can fix that, if we want).

This can add up to a great overhead with a large number of RPC
connections. Users of seastar should consider enabling it selectively,
only for the most voluminous connections.

Scylla, for example, could enable it only for connections to sibling
shard on other nodes, but not for non-sibling shards on other nodes.
This would reap near-full benefits of compression in well-behaved
(i.e. homogeneous) clusters, without inviting a catastrophical buffer
size explosion in large clusters.