build: disable implicit fallthough
this change passes "-Wimplicit-fallthrough" compiling option to
the compiler, so it warns at seeing implicit fallthrough. and
since we take all warnings as errors, with the exception of
"-Wdeprecated-declaration", this warning is taken as error.

so we will prevent switch-case statement from fallingh through
without being annotated with the C++ attribute of "[[fallthough]]".

this should help us to identify the potential programming bugs
at compile time.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>