circular_buffer_fixed_capacity_test: enable randomize test to reverse
before this change, the randomized test accepts a RNG which generates
numbers in the range of [0, 6], and translate the generated numbers
to corresponding operations, then apply them to the `std::deque` and
`circular_buffer_fixed_capacity<>` under test.

it seems that we intent to support numbers in the range of [0, 7],
as we also handle 7, and reverse the dequeu and buffer respectively.
but we failed generate 7 with the RNG.

so, in this change,

* enlarge the maximum value of of the uniform distribution from 6 to 7
* do not fallthrough in the "case 7" block, otherwise the test will
  abort() at seeing "7" generated from RNG.

this change should improve the coverage of the test. and enable us to
pass `-Wimplicit-fallthrough` command line option to the compiler for
better compile-time check.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>